product page changed so links only apply to the allowed users, controller methods/views shifted from profiles to carts and wishlists
