+++ app/assets/stylesheets/application.scss
@@ -76,7 +76,7 @@
 }
 
 a {
-    color: white;
+    color: black;   
 }
 p {
     color: black;
