+++ app/assets/stylesheets/products.scss
@@ -133,3 +133,15 @@
     margin-left: 10rem;
 }
 
+
+.add-product-btn {
+    background-color: #4CAF50; /* Green */
+    border: none;
+    color: white;
+    padding: 15px 32px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 16px;
+}
+
