+++ app/assets/stylesheets/profiles.scss
@@ -103,4 +103,3 @@
 #out-of-stock-notice {
     color: #ff3333;
 }
-
