+++ app/helpers/wishlists_helper.rb
@@ -1,0 +1,2 @@
+module WishlistsHelper
+end
