+++ app/views/layouts/application.html.erb
@@ -50,7 +50,7 @@
               
                 <% if current_user.profile.present? %>
                     <div class="cart-icon">
-                        <%= link_to image_tag('shopping-cart-grey.svg'), {:controller => "profiles", :action => "showcart", :id => current_user.profile.id } %>
+                        <%= link_to image_tag('shopping-cart-grey.svg'), {:controller => "carts", :action => "showcart", :id => current_user.profile.id } %>
                     </div>
                 <% end %>
