+++ app/views/products/show.html.erb
@@ -3,17 +3,26 @@
     <% if notice == 'The item has been successfully added to your cart.' %>
       <div class="message-wrap">
         <%= inline_svg("_ionicons_svg_md-checkmark-circle.svg", class: "check-icon") %>
-        <p id="item-added-notice"><%= notice %></p>    <%# need to have a notice that pops up saying that the item was successfully added %>
+        <p id="item-added-notice"><%= notice %></p>
         <p> <%= link_to 'To cart.', "/profiles/#{current_user.profile.id}/showcart" %> </p>
-        <%= link_to inline_svg("_ionicons_svg_md-close.svg", class: "exit-icon"), "/products/#{@product.id}?fit=#{params[:fit]}", class: 'exit-link' %>
-
+        <%= link_to inline_svg("_ionicons_svg_md-close.svg", class: "exit-icon"), "/products/#{@product.id}?product_size=#{params[:product_size]}", class: 'exit-link' %>
       </div>
-    <%# might not need this anymore - NO I STILL WILL IF THE USER ADDS ANOTHER ONE  %>
-
-    <% elsif notice == 'Sorry, you have ordered more items than there are in stock.' %>
+    <% elsif notice == 'Sorry, you need to sign in to add to cart.' %>
       <div class="message-wrap">
-        <p id="qty-exceeded-notice"><%= notice %></p> 
+        <p id="qty-exceeded-notice"><%= notice %></p>
+        <%= link_to 'Sign in?', "/users/sign_in", class: 'exit-link' %> 
         <%= link_to inline_svg("_ionicons_svg_md-close.svg", class: "exit-icon"), "/products/#{@product.id}", class: 'exit-link' %>
+      </div>
+    <% elsif notice == 'Sorry, you need to sign in to add to wishlist.' %>
+      <div class="message-wrap">
+        <p id="qty-exceeded-notice"><%= notice %></p>
+        <%= link_to 'Sign in?', "/users/sign_in", class: 'exit-link' %> 
+        <%= link_to inline_svg("_ionicons_svg_md-close.svg", class: "exit-icon"), "/products/#{@product.id}", class: 'exit-link' %>
+      </div>
+    <% elsif (notice != nil) && (notice.include? "out of stock") %>
+      <div class="message-wrap">
+        <p id="out-of-stock-notice"><%= notice %></p>
+        <%= link_to inline_svg("_ionicons_svg_md-close.svg", class: "exit-icon"), "/products/#{@product.id}?product_size=#{params[:product_size]}", class: 'exit-link' %>
       </div>
     <% end %>
     
@@ -24,7 +33,7 @@
           <%= image_tag @product.uploaded_image.service_url, class: 'image-size' %>
       </div>
     <% else %>
-      <p>No image for this product</p>
+      <p>There were errors displaying the image</p>
     <% end %>
 
     <div class="product-info">
@@ -41,9 +50,7 @@
         </p>
 
         <p>
-
           <%= link_to @product.category.category, "/#{@product.category}" %>
-
         </p>
 
         <p>
@@ -51,31 +58,26 @@
         </p>
 
         <p>
-          <% if params[:fit].nil? %>
-            <%= "$#{@product.product_variants.order(:size_id).first.price}" %>
-            <% params[:fit] = @product.product_variants.order(:size_id).first.size_id %>
-          <% elsif %>
-            <%= "$#{@product.product_variants.find_by(size_id: params[:fit]).price}" %>
-          <% end %>
+          <%= "$#{price(params[:product_size], @product)}" %>
         </p>
+        <% params[:product_size] = set_product_size(params[:product_size], @product) %>
+
 
         <div class="add-to">
-          <% if user_signed_in? && current_user.profile != @product.profile %>
-            <%= form_with(url: "/profiles/#{current_user.profile.id}/addproducttocart?fit=#{params[:fit]}", local: true) do |form| %>
-
-              <% possible_sizes = [] %>
-
-              <% @product.product_variants.each do |variant| %>
-                <% possible_sizes << variant.size_id %>
-              <% end %>
-
+          <% if !user_signed_in? || (user_signed_in? && !owning_artist(@product)) %>
+            <% if !user_signed_in? %>
+              <% form_url_path = "/user_does_not_exist?product_id=#{@product.id}&product_size=#{params[:product_size]}" %>
+            <% elsif user_signed_in? && !owning_artist(@product) %>
+              <% form_url_path = "/profiles/#{current_user.profile.id}/addproducttocart?product_size=#{params[:product_size]}" %>
+            <% end %>
+            
+            <%= form_with(url: form_url_path, local: true) do |form| %>
               <div class="field" >
-                  <%= form.label :size %>
-                  <%= form.collection_select :size_id, Size.where(category: @product.category.category).where(id: possible_sizes), :id, :sizing, {
-                      :selected => params[:fit] }, { :onchange => "window.location.replace('/products/#{@product.id}?fit='+this.value);" } %>
+                <%= form.label :size %>
+                <%= form.collection_select :size_id, Size.where(category: @product.category.category).where(id: possible_sizes(@product)), :id, :sizing, {
+                      :selected => params[:product_size] }, { :onchange => "window.location.replace('/products/#{@product.id}?product_size='+this.value);" } %>
               </div>
 
-            
               <div class="field">
                 <%= form.hidden_field :qty, value: 1 %>
               </div>
@@ -85,43 +87,31 @@
               </div>
               
               <div class="actions">
-                <%= form.submit "Add to cart", class: "button"%>
+                <%= form.submit "Add to cart", class: "add-product-btn"%>
               </div>
-
             <% end %>
           <% end %>
         </div>
 
     </div>
 
-    
     <div class="wishlist-box">
-      <% if current_user.profile.wishlist.products.where("product_id = ?", @product.id).present? %>
-          <%= link_to inline_svg("_ionicons_svg_md-star.svg", class: "wishlist-icon-active"), "/profiles/#{current_user.profile.id}/show-wishlist/remove-item?product_id=#{@product.id}" %>
-      <% else %>
-          <%= link_to inline_svg("_ionicons_svg_md-star.svg", class: "wishlist-icon-inactive"), "/profiles/#{current_user.profile.id}/addtowishlist?product_id=#{@product.id}" %>
-      <% end %>
-    </div>
-
-      <% if user_signed_in? %>
-        <% if current_user.profile != nil && current_user.profile != @product.profile %>
-          <% if current_user.profile == @product.profile %> 
-                <%= link_to 'Delete', "/products/#{@product.id}", method: :delete %> |
-                <%= link_to 'Edit', edit_product_path(@product) %> |
-                <%= link_to 'back to profile', "/profiles/#{@product.profile.id}" %>
+      <% if !user_signed_in? %>
+        <%= link_to inline_svg("_ionicons_svg_md-star.svg", class: "wishlist-icon-inactive"), "/user_does_not_exist_wishlist?product_id=#{@product.id}&product_size=#{params[:product_size]}" %>
+      <% elsif user_signed_in? %>
+        <% if !owning_artist(@product) %>
+          <% if current_user.profile.wishlist.products.where("product_id = ?", @product.id).present? %>
+            <%= link_to inline_svg("_ionicons_svg_md-star.svg", class: "wishlist-icon-active"), "/profiles/#{current_user.profile.id}/show-wishlist/remove-item?product_id=#{@product.id}" %>
+          <% else %>
+            <%= link_to inline_svg("_ionicons_svg_md-star.svg", class: "wishlist-icon-inactive"), "/profiles/#{current_user.profile.id}/addtowishlist?product_id=#{@product.id}" %>
           <% end %>
-
-
-        <% else %>  <%#user needs to be signed in in order to add the product to their cart or their wishlist %>
-          <div class="add-to-cart">
-            <%= link_to 'Add to cart', "/users/sign_in", data: { confirm: 'You must log in to add to cart, are you sure?' } %>
-          </div>
-          <%= link_to 'Add to wishlist', "/users/sign_in", data: { confirm: 'You must log in to add to wishlist, are you sure?' }%>
-       
+        <% elsif owning_artist(@product) %>
+          <%#show nada%>
         <% end %>
-      <% end %>
+      <% end %>    
+    </div>
     
-    <% if user_signed_in? && current_user.profile == @product.profile %>
+    <% if user_signed_in? && owning_artist(@product) %>
 
       <br><br><br>
