+++ app/views/wishlists/show_wishlist.html.erb
@@ -20,7 +20,7 @@
                      
               
                     <div class="remove">
-                         <%= link_to 'Remove Item', "/profiles/#{params[:id]}/show-wishlist/remove-item?product_id=#{product.id}" %>
+                         <%= link_to 'Remove Item', "/profiles/#{params[:id]}/show-wishlist/remove-item-on-wishlist?product_id=#{product.id}" %>
                     </div>
