+++ config/routes.rb
@@ -5,34 +5,35 @@ Rails.application.routes.draw do
   devise_for :users
   root to: 'pages#fandoms'
 
+  #PROFILES
+  post '/user_does_not_exist' => 'profiles#no_user'
+  get '/user_does_not_exist_wishlist' => 'profiles#no_user_for_wishlist'
 
+
+  #PRODUCTS
   get '/products/:id/add-variant' => 'product_variants#new'
-
   get '/products/:product_id/edit-variant/:id' => 'product_variants#edit'
   get '/products/:product_id/remove-variant/:id' => 'product_variants#destroy'
 
 
-  post '/profiles/:id/addproducttocart' => 'profiles#add_product_to_cart'
+  #CART
+  get '/profiles/:id/addtocart' => 'carts#add_to_cart' 
+  get '/profiles/:id/showcart' => 'carts#showcart' 
+  get '/profiles/:id/showcart/remove-item' => 'carts#remove_cart_item'
+  get '/profiles/:id/checkout_cart' => 'carts#checkout_cart'
+  get '/profiles/:id/previous_orders' => 'carts#previous_orders'
+  post '/profiles/:id/addproducttocart' => 'carts#add_product_to_cart'   
+  get '/profiles/:id/showcart/:purchase_id/change-qty' => 'carts#change_item_qty' 
 
-  get '/profiles/:id/showcart/:purchase_id/change-qty' => 'profiles#change_item_qty'   #need to find a better way to do this through the controller
 
-
-  get '/profiles/:id/addtocart' => 'profiles#add_to_cart' 
-  get '/profiles/:id/showcart' => 'profiles#showcart'
-  get '/profiles/:id/showcart/remove-item' => 'profiles#remove_cart_item'
-  get '/profiles/:id/checkout_cart' => 'profiles#checkout_cart'
-  get '/profiles/:id/previous_orders' => 'profiles#previous_orders'
-
-  #wishlist stuff
-  get '/profiles/:id/addtowishlist' => 'profiles#add_to_wishlist'
-  get '/profiles/:id/show-wishlist' => 'profiles#show_wishlist'
-  get '/profiles/:id/show-wishlist/remove-item' => 'profiles#remove_wishlist_item'
-  get '/profiles/:id/show-wishlist/remove-item-on-wishlist' => 'profiles#remove_item_on_wishlist_page'
-
+  #WISHLIST
+  get '/profiles/:id/addtowishlist' => 'wishlists#add_to_wishlist' 
+  get '/profiles/:id/show-wishlist' => 'wishlists#show_wishlist' 
+  get '/profiles/:id/show-wishlist/remove-item' => 'wishlists#remove_wishlist_item'
+  get '/profiles/:id/show-wishlist/remove-item-on-wishlist' => 'wishlists#remove_item_on_wishlist_page' 
 
 
   # categories and fandoms
-
   get '/categories/:id' => 'pages#category'
   get '/tshirts' => 'pages#tshirts'
   get '/iphonecases' => 'pages#iphonecases'
@@ -40,23 +41,8 @@ Rails.application.routes.draw do
   get '/prints' => 'pages#prints'
 
 
-
-
   #Fandom pages "oh what fun..."
   get '/fandoms/:id' => 'fandoms#fandom'
-
-  # get '/marvel' => 'fandoms#marvel'
-  # get '/dc' => 'fandoms#dc'
-  # get '/star_wars' => 'fandoms#star_wars'
-  # get '/dr_who' => 'fandoms#dr_who'
-  # get '/the_walking_dead' => 'fandoms#the_walking_dead'
-  # get '/supernatural' => 'fandoms#supernatural'
-  # get '/pokemon' => 'fandoms#pokemon'
-  # get '/dragonball_z' => 'fandoms#dragonball_z'
-  # get '/fullmetal_alchemist' => 'fandoms#fullmetal_alchemist'
-  # get '/sherlock' => 'fandoms#sherlock'
-  # get '/adventure_time' => 'fandoms#adventure_time'
-  # get '/game_of_thrones' => 'fandoms#game_of_thrones'
 
   resources :charges
   get 'profiles/:id/become_artist' => 'profiles#become_an_artist'
